requirements.txt
requests
terminaltables
pyzmq
flask
termcolor
beautifulsoup4
websocket-client
ffmpy
m3u8
fake-useragent
django-environ
pycurl
pycurl-requests
python-socketio==4.6.1